Fixes #2096 spatialFilter redraw
