+++ web/client/epics/wfsquery.js
@@ -15,6 +15,9 @@ const {FEATURE_TYPE_SELECTED, QUERY, featureLoading, featureTypeLoaded, featureT
 const {paginationInfo, isDescribeLoaded, describeSelector} = require('../selectors/query');
 const FilterUtils = require('../utils/FilterUtils');
 const assign = require('object-assign');
+const {spatialFieldSelector, spatialFieldGeomTypeSelector, spatialFieldGeomCoordSelector, spatialFieldGeomSelector, spatialFieldGeomProjSelector} = require('../selectors/queryform');
+const {changeDrawingStatus, CHANGE_DRAWING_STATUS} = require('../actions/draw');
+const {SET_CONTROL_PROPERTY} = require('../actions/controls');
 
 const {isObject} = require('lodash');
 const {interceptOGCError} = require('../utils/ObservableUtils');
@@ -283,15 +286,35 @@ const viewportSelectedEpic = (action$, store) =>
             return Rx.Observable.empty();
         });
 
+const redrawSpatialFilterEpic = (action$, store) => action$.ofType(SET_CONTROL_PROPERTY)
+    .filter(action => action.control && action.control === 'queryPanel' && action.value)
+    .switchMap(() => {
+        const state = store.getState();
+        const spatialField = spatialFieldSelector(state);
+        const feature = {
+            type: "Feature",
+            geometry: {
+                type: spatialFieldGeomTypeSelector(state),
+                coordinates: spatialFieldGeomCoordSelector(state)
+            }
+        };
+        let drawSpatialFilter = spatialFieldGeomSelector(state) ? changeDrawingStatus("drawOrEdit", spatialField.method, "queryform", [feature], {featureProjection: spatialFieldGeomProjSelector(state), drawEnabled: false, editEnabled: false}) : changeDrawingStatus("clean", spatialField.method, "queryform", [], {drawEnabled: false, editEnabled: false});
+         // if a geometry is present it will redraw the spatial field
+        return Rx.Observable.of(drawSpatialFilter);
+    });
+
+
  /**
   * Epics for WFS query requests
   * @name epics.wfsquery
   * @type {Object}
   */
 
+
 module.exports = {
     featureTypeSelectedEpic,
     wfsQueryEpic,
+    redrawSpatialFilterEpic,
     viewportSelectedEpic,
     getWFSFilterData
 };
